+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/TicketSchemes/TicketScheme.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 
 public interface TicketScheme {
 
-    Scored<BigDecimal> getBestAmount();
+    Scored<BigDecimal> getBestAmount(boolean strict);
 
     String toString();
 }