+++ ocr/src/main/java/com/ing/software/ocr/OperativeObjects/RawImage.java
@@ -5,16 +5,23 @@ import android.graphics.RectF;
 import android.support.annotation.NonNull;
 
 import com.annimon.stream.Stream;
+import com.ing.software.ocr.OcrManager;
 import com.ing.software.ocr.OcrObjects.TempText;
 import com.ing.software.ocr.OcrUtils;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import static com.ing.software.ocr.OcrVars.*;
 
 /**
- * Class to store only useful properties of source images
+ * Class to store only useful properties of source images and scheme of ticket
+ * How to use:
+ * - initialize with constructor
+ * - call setLines() passing a list of all TempTexts
+ * - call OcrSchemer.prepareScheme()
+ * - now you can use the others methods in this class
  * @author Michelon
  */
 
@@ -109,11 +116,12 @@ public class RawImage {
      * @return list of Intro RawTexts
      */
     public List<TempText> getIntroTexts() {
+        Collections.sort(introTexts);
         return introTexts;
     }
 
     public RectF getIntroRect() {
-        return introRect;
+        return introRect != null? introRect : new RectF(0, 0, width, 0);
     }
 
     /**
@@ -121,11 +129,12 @@ public class RawImage {
      * @return list of Products RawTexts
      */
     public List<TempText> getProductsTexts() {
+        Collections.sort(productsTexts);
         return productsTexts;
     }
 
     public RectF getProductsRect() {
-        return productsRect;
+        return productsRect != null? productsRect : new RectF(0, getIntroRect().bottom, width/2, getConclusionRect().top);
     }
 
     /**
@@ -133,11 +142,12 @@ public class RawImage {
      * @return list of Prices RawTexts
      */
     public List<TempText> getPricesTexts() {
+        Collections.sort(pricesTexts);
         return pricesTexts;
     }
 
     public RectF getPricesRect() {
-        return pricesRect;
+        return pricesRect != null? pricesRect : new RectF(width/2, getIntroRect().bottom, width, getConclusionRect().top);
     }
 
     /**
@@ -145,11 +155,12 @@ public class RawImage {
      * @return list of Conclusion RawTexts
      */
     public List<TempText> getConclusionTexts() {
+        Collections.sort(conclusionTexts);
         return conclusionTexts;
     }
 
     public RectF getConclusionRect() {
-        return conclusionRect;
+        return conclusionRect != null? conclusionRect : new RectF(0, height, width, height);
     }
 
     /**
@@ -159,7 +170,7 @@ public class RawImage {
     private float checkAverageLineHeight() {
         float average = Stream.of(allTexts)
                 .reduce(0f, (sum, currentLine) -> sum + currentLine.height()) / allTexts.size();
-        OcrUtils.log(3, "RawImage", "AVERAGE RECT HEIGHT is: " + average);
+        OcrUtils.log(5, "RawImage", "AVERAGE RECT HEIGHT is: " + average);
         return average;
     }
 
@@ -170,7 +181,7 @@ public class RawImage {
     private float checkAverageCharHeight() {
         float average = Stream.of(allTexts)
                 .reduce(0f, (sum, currentLine) -> sum + currentLine.charHeight()) / allTexts.size();
-        OcrUtils.log(3, "RawImage", "AVERAGE CHAR HEIGHT is: " + average);
+        OcrUtils.log(5, "RawImage", "AVERAGE CHAR HEIGHT is: " + average);
         return average;
     }
 
@@ -181,7 +192,7 @@ public class RawImage {
     private float checkAverageCharWidth() {
         float average = Stream.of(allTexts)
                 .reduce(0f, (sum, currentLine) -> sum + currentLine.charWidth()) / allTexts.size();
-        OcrUtils.log(3, "RawImage", "AVERAGE CHAR WIDTH is: " + average);
+        OcrUtils.log(5, "RawImage", "AVERAGE CHAR WIDTH is: " + average);
         return average;
     }
 
@@ -190,6 +201,11 @@ public class RawImage {
      * Must be used only once and after setLines() has been called and tags have been set.
      */
     public void textFitter() {
+        //Remove old lists
+        introTexts = new ArrayList<>();
+        pricesTexts = new ArrayList<>();
+        productsTexts = new ArrayList<>();
+        conclusionTexts = new ArrayList<>();
         for (TempText text : allTexts) {
             if (text.getTags().contains(INTRODUCTION_TAG))
                 introTexts.add(text);
@@ -210,14 +226,14 @@ public class RawImage {
      * Get rect containing all lines detected
      * @return Rect containing all rects passed
      */
-    private RectF getMaxRectBorders(List<TempText> allTexts) {
-        if (allTexts.size() == 0)
+    private RectF getMaxRectBorders(List<TempText> texts) {
+        if (texts.size() == 0)
             return null;
         float left = width;
         float right = 0;
         float top = height;
         float bottom = 0;
-        for (TempText text : allTexts) {
+        for (TempText text : texts) {
             if (text.box().left < left)
                 left = text.box().left;
             if (text.box().right > right)
@@ -228,5 +244,127 @@ public class RawImage {
                 top = text.box().top;
         }
         return new RectF(left, top, right, bottom);
+    }
+
+    /**
+     * Override a text with a new one
+     * @param oldText
+     * @param newText
+     */
+    public void overrideText(TempText oldText, TempText newText) {
+        allTexts.remove(oldText);
+        allTexts.add(newText);
+        //fail silently
+        introTexts.remove(oldText);
+        pricesTexts.remove(oldText);
+        productsTexts.remove(oldText);
+        conclusionTexts.remove(oldText);
+        if (newText.getTags().contains(INTRODUCTION_TAG)) {
+            introTexts.add(newText);
+            introRect = getMaxRectBorders(introTexts);
+        } else if (newText.getTags().contains(PRODUCTS_TAG)) {
+            productsTexts.add(newText);
+            productsRect = getMaxRectBorders(productsTexts);
+        } else if (newText.getTags().contains(PRICES_TAG)) {
+            pricesTexts.add(newText);
+            pricesRect = getMaxRectBorders(pricesTexts);
+        } else if (newText.getTags().contains(CONCLUSION_TAG)) {
+            conclusionTexts.add(newText);
+            conclusionRect = getMaxRectBorders(conclusionTexts);
+        }
+    }
+
+    /**
+     * Remove a text from all lists
+     * @param oldText
+     */
+    public void removeText(TempText oldText) {
+        allTexts.remove(oldText);
+        //fail silently
+        introTexts.remove(oldText);
+        pricesTexts.remove(oldText);
+        productsTexts.remove(oldText);
+        conclusionTexts.remove(oldText);
+        introRect = getMaxRectBorders(introTexts);
+        productsRect = getMaxRectBorders(productsTexts);
+        pricesRect = getMaxRectBorders(pricesTexts);
+        conclusionRect = getMaxRectBorders(conclusionTexts);
+    }
+
+    /**
+     * Remove all texts inside passed rect from all lists
+     * @param rect
+     */
+    public void removeText(RectF rect) {
+        List<TempText> tempList = new ArrayList<>(allTexts);
+        for (TempText text : tempList) {
+            if (rect.contains(text.box())) {
+                allTexts.remove(text);
+                introTexts.remove(text);
+                pricesTexts.remove(text);
+                productsTexts.remove(text);
+                conclusionTexts.remove(text);
+            }
+        }
+        //fail silently
+        introRect = getMaxRectBorders(introTexts);
+        productsRect = getMaxRectBorders(productsTexts);
+        pricesRect = getMaxRectBorders(pricesTexts);
+        conclusionRect = getMaxRectBorders(conclusionTexts);
+    }
+
+    /**
+     * Add text to lists
+     * @param newText
+     */
+    public void addText(TempText newText) {
+        allTexts.add(newText);
+        if (newText.getTags().contains(INTRODUCTION_TAG)) {
+            introTexts.add(newText);
+            introRect = getMaxRectBorders(introTexts);
+        } else if (newText.getTags().contains(PRODUCTS_TAG)) {
+            productsTexts.add(newText);
+            productsRect = getMaxRectBorders(productsTexts);
+        } else if (newText.getTags().contains(PRICES_TAG)) {
+            pricesTexts.add(newText);
+            pricesRect = getMaxRectBorders(pricesTexts);
+        } else if (newText.getTags().contains(CONCLUSION_TAG)) {
+            conclusionTexts.add(newText);
+            conclusionRect = getMaxRectBorders(conclusionTexts);
+        }
+    }
+
+    /**
+     * Add tag to text according to its position (must call textFitter before)
+     * @param text
+     * @return
+     */
+    public static TempText mapText(TempText text) {
+        if (OcrManager.mainImage.getIntroRect().contains(text.box()))
+            text.addTag(INTRODUCTION_TAG);
+        else if (OcrManager.mainImage.getProductsRect().contains(text.box()))
+            text.addTag(PRODUCTS_TAG);
+        else if (OcrManager.mainImage.getPricesRect().contains(text.box()))
+            text.addTag(PRICES_TAG);
+        else if (OcrManager.mainImage.getConclusionRect().contains(text.box()))
+            text.addTag(CONCLUSION_TAG);
+        else {
+            float productsTop = Math.min(OcrManager.mainImage.getProductsRect().top, OcrManager.mainImage.getPricesRect().top);
+            float productsBottom = Math.max(OcrManager.mainImage.getProductsRect().bottom, OcrManager.mainImage.getPricesRect().bottom);
+            RectF middleRect = new RectF(OcrManager.mainImage.getProductsRect().left, productsTop, OcrManager.mainImage.getPricesRect().right, productsBottom);
+            //Check its y coordinate
+            float centerY = text.box().centerY();
+            if (centerY < productsTop)
+                text.addTag(INTRODUCTION_TAG);
+            else if (centerY > productsBottom)
+                text.addTag(CONCLUSION_TAG);
+            else {
+                if (text.box().centerX() < middleRect.width()/2)
+                    text.addTag(PRODUCTS_TAG);
+                else
+                    text.addTag(PRICES_TAG);
+            }
+        }
+        return text;
     }
 }